Prepare parser for serde implementation
