+++ parser/src/lib.rs
@@ -1,6 +1,6 @@
-mod types;
-
+#![forbid(unsafe_code)]
 pub mod load;
 pub mod parse;
+pub mod types;
 
 pub use types::{Component, HailValue, Matrix, Table};
