+++ parser/src/parse/data/mod.rs
@@ -4,6 +4,8 @@ mod error;
 mod parse;
 mod parse_rows;
 
+pub mod helpers;
+
 pub use encoders::{Encoding, StandardEncoder, UnsignedLEB128Encoder};
 pub use error::DataParsingError;
 pub use parse_rows::parse_rows;
