+++ parser/src/parse/mod.rs
@@ -1,7 +1,8 @@
-mod data;
 pub(crate) mod schema;
 
-pub use data::parse_rows;
+pub(crate) use data::parse_rows;
+
+pub mod data;
 pub use data::{Encoding, StandardEncoder, UnsignedLEB128Encoder};
 
 pub use data::DataParsingError;
