+++ parser/tests/tests.rs
@@ -1,0 +1,3 @@
+mod component;
+mod matrix_tables;
+mod table;
